+++ changelog.md
@@ -1,2 +1,7 @@
-## 21.1.1
-- Fixed using wrong jar task for publish mod
+## 21.1.2
+### NeoForge
+- Added compatibility to menu that use `DisableableSlot` from `Titanium`
+- Added compatibility to gui that wrap player inventory to `InvWrapper`
+
+### Fabric
+No changes.